refactor: move networks to icp-eth modules
