+++ src/frontend/src/eth/components/send/SendNetwork.svelte
@@ -7,8 +7,7 @@
 	import { ckEthHelperContractAddressStore } from '$icp-eth/stores/cketh.store';
 	import { getContext } from 'svelte';
 	import { SEND_CONTEXT_KEY, type SendContext } from '$icp-eth/stores/send.store';
-	import { ICP_NETWORK } from '$icp/constants/networks.constants';
-	import { ETHEREUM_NETWORK } from '$eth/constants/networks.constants';
+	import { ETHEREUM_NETWORK, ICP_NETWORK } from '$icp-eth/constants/networks.constants';
 
 	export let network: Network | undefined = undefined;
 	export let destination: string | undefined = undefined;
