+++ src/frontend/src/eth/services/fee.services.ts
@@ -1,11 +1,11 @@
 import { CKETH_FEE } from '$eth/constants/cketh.constants';
 import { ERC20_FALLBACK_FEE } from '$eth/constants/erc20.constants';
 import { ETH_BASE_FEE } from '$eth/constants/eth.constants';
-import { ETHEREUM_NETWORK } from '$eth/constants/networks.constants';
 import { getFeeData as getBurnFeeData } from '$eth/providers/infura-erc20-icp.providers';
 import { getFeeData as getErc20FeeDataProvider } from '$eth/providers/infura-erc20.providers';
 import type { Erc20ContractAddress } from '$eth/types/erc20';
 import { isCkEthHelperContract } from '$eth/utils/send.utils';
+import { ETHEREUM_NETWORK } from '$icp-eth/constants/networks.constants';
 import type { CkEthHelperContractAddressData } from '$icp-eth/stores/cketh.store';
 import type { ETH_ADDRESS } from '$lib/types/address';
 import type { Network } from '$lib/types/network';
