+++ src/frontend/src/icp-eth/constants/networks.constants.ts
@@ -1,6 +1,10 @@
 import eth from '$icp-eth/assets/eth.svg';
+import icpLight from '$icp/assets/icp_light.svg';
 import type { Network } from '$lib/types/network';
 
+/**
+ * Ethereum
+ */
 export const ETHEREUM_NETWORK_SYMBOL = import.meta.env.VITE_ETHEREUM_SYMBOL;
 
 export const ETHEREUM_NETWORK_ID = Symbol(ETHEREUM_NETWORK_SYMBOL);
@@ -9,4 +13,17 @@ export const ETHEREUM_NETWORK: Network = {
 	id: ETHEREUM_NETWORK_ID,
 	name: 'Ethereum',
 	icon: eth
+};
+
+/**
+ * ICP
+ */
+export const ICP_NETWORK_SYMBOL = 'ICP';
+
+export const ICP_NETWORK_ID = Symbol(ICP_NETWORK_SYMBOL);
+
+export const ICP_NETWORK: Network = {
+	id: ICP_NETWORK_ID,
+	name: 'Internet Computer',
+	icon: icpLight
 };
