+++ src/frontend/src/icp/components/convert/HowToConvertETHModal.svelte
@@ -8,7 +8,7 @@
 	import ConvertETHToCkETHWizard from '$icp-eth/components/send/ConvertETHToCkETHWizard.svelte';
 	import { HOW_TO_CONVERT_WIZARD_STEPS } from '$icp-eth/constants/how-to-convert.constants';
 	import { closeModal } from '$lib/utils/modal.utils';
-	import { ICP_NETWORK } from '$icp/constants/networks.constants';
+	import { ICP_NETWORK } from '$icp-eth/constants/networks.constants';
 
 	/**
 	 * Props
