+++ src/frontend/src/icp/components/send/IcNetwork.svelte
@@ -6,7 +6,7 @@
 	import { isNetworkIdBTC } from '$icp/utils/ic-send.utils';
 	import { isNetworkIdEthereum } from '$lib/utils/network.utils';
 
-	import { ETHEREUM_NETWORK } from '$eth/constants/networks.constants';
+	import { ETHEREUM_NETWORK } from '$icp-eth/constants/networks.constants';
 
 	export let networkId: NetworkId | undefined = undefined;
 </script>
