+++ src/frontend/src/icp/components/send/IcSendModal.svelte
@@ -24,7 +24,8 @@
 	import BitcoinFeeContext from '$icp/components/fee/IcFeeContext.svelte';
 	import { BTC_NETWORK_ID } from '$icp/constants/ckbtc.constants';
 	import { closeModal } from '$lib/utils/modal.utils';
-	import { ETHEREUM_NETWORK_ID } from '$eth/constants/networks.constants';
+
+	import { ETHEREUM_NETWORK_ID } from '$icp-eth/constants/networks.constants';
 
 	/**
 	 * Props
