+++ src/frontend/src/lib/constants/networks.constants.ts
@@ -1,5 +1,4 @@
-import { ETHEREUM_NETWORK } from '$eth/constants/networks.constants';
-import { ICP_NETWORK } from '$icp/constants/networks.constants';
+import { ETHEREUM_NETWORK, ICP_NETWORK } from '$icp-eth/constants/networks.constants';
 import type { Network } from '$lib/types/network';
 
 export const NETWORKS: Network[] = [ETHEREUM_NETWORK, ICP_NETWORK];
