+++ src/frontend/src/lib/constants/tokens.constants.ts
@@ -1,12 +1,11 @@
-import { ETHEREUM_NETWORK } from '$eth/constants/networks.constants';
 import eth from '$icp-eth/assets/eth.svg';
+import { ETHEREUM_NETWORK, ICP_NETWORK } from '$icp-eth/constants/networks.constants';
 import icpLight from '$icp/assets/icp_light.svg';
 import {
 	ICP_INDEX_CANISTER_ID,
 	ICP_LEDGER_CANISTER_ID,
 	ICP_TRANSACTION_FEE_E8S
 } from '$icp/constants/icp.constants';
-import { ICP_NETWORK } from '$icp/constants/networks.constants';
 import type { IcToken } from '$icp/types/ic';
 import type { Token } from '$lib/types/token';
